update modal style
