+++ assets/css/tablet.scss
@@ -545,64 +545,9 @@
   }
 
   .modal {
-    position: fixed;
-    left: 50%;
-    top: 50%;
-    transform: translate(-50%, -50%);
-    // display: flex;
-    // flex-direction: column;
-    // justify-content: center;
-    //  align-items: center;
-    // width: 600px;
-    // min-width: 90%;
     z-index: 10;
-    // background: yellow;
   }
-
-  .modalBackground {
-    position: fixed;
-    left: 50%;
-    top: 50%;
-    transform: translate(-50%, -50%);
-    // display: flex;
-    // position: fixed;
-    // top: 0;
-    justify-content: center;
-    // align-items: center;
-    // position: absolute;
-    // background: rgba(193, 199, 208, 0.8);
-    // backdrop-filter: blur(10px);
-    width: 90%;
-    height: 80%;
-    z-index: 15;
-    background: white;
-    border-radius: 8px;
-    // margin: 10%;
-    padding: 2%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 1rem;
-  }
-
-  .modal-body {
-    /*    background: white;
-    border-radius: 8px;
-    margin: 10%;
-    padding: 2%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 1rem; */
-    // position: absolute;
-    // top: 0;
-  }
-
-  .modal-container {
-    display: flex;
-    flex-direction: column;
-  }
-
+  
   .modal-list {
     display: flex;
     gap: 8px;
