+++ index.html
@@ -15,7 +15,7 @@
     <header class="header-menu">
         <nav class="nav-bar">
                 <img src="./assets/img/hamburguer.svg" alt="icon" id="hamburg-icon" class="hamburguer">
-                <ul class="nav-links hide">
+                <ul class="nav-links">
                 <li ><a href="#work-section-area" id="item-portfolio">Portfolio</a></li>
                 <li ><a href="#about-section-area" id="item-about">About</a></li>
                 <li ><a href="#contact-section-area" id="item-contact">Contact</a></li>
@@ -381,7 +381,6 @@
     <div class="fade hide"></div>
     <div class="modal" id="modal">
     </div>
-    <script src="./assets/js/main.js"></script>
-    <script src="./assets/js/desktop.js"></script>
+    <script type="module" src="./assets/js/main.js"></script>
 </body>
 </html>
